--- client/src/components/faculty/Dashboard.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Dashboard = () => {
-  return (
-    <div>
-      <h1>Faculty Dashboard</h1>
-    </div>
-  );
-};
-
-export default Dashboard;
