--- client/src/components/student/Dashboard.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Dashboard = () => {
-  return (
-    <div>
-      <h1>Student Dashboard</h1>
-    </div>
-  );
-};
-
-export default Dashboard;
