+++ client/src/components/student/Student.jsx
@@ -2,8 +2,7 @@ import React from "react";
 import { Switch, Route } from "react-router-dom";
 import Login from "../common/Login";
 import Register from "./auth/Register";
-// import Dashboard from "./Dashboard";
-import ProtectedRoute from "../../components/ProtectedRoute";
+import ProtectedRoute from "../../components/common/ProtectedRoute";
 import ApplicationsList from "../applications/ApplicationsList";
 import ApplicationDetail from "../applications/ApplicationDetail";
 import StudentActions from "../applications/StudentActions";
@@ -16,12 +15,6 @@ const Student = () => {
         <Login userType="student" name="Student" />
       </Route>
       <Route path="/student/register" component={Register} />
-      {/* <ProtectedRoute
-        exact
-        path="/student"
-        component={Dashboard}
-        userType={"student"}
-      /> */}
       <ProtectedRoute
         exact
         path="/student/profile"
